address review comments (split tests, test for no vari on stack when double only case)
